Move X event handling logic into the x module
